reduce scope to only LIQUID and small refactor
