+++ brambl-sdk/src/main/scala/co/topl/brambl/builders/TransactionBuilderApi.scala
@@ -535,6 +535,7 @@ object TransactionBuilderApi {
 
       // We cannot split an IMMUTABLE or ACCUMULATOR quantity type, but due to user validation, if inValue is one of
       // those 2 types, then changeQuantity must be 0.
+      // TODO: Currently we are only supporting LIQUID quantity descriptor type
       private def buildChange(inValue: BoxValue, changeQuantity: Int128): Option[BoxValue] =
         // User validation guarantees that the following will return None for IMMUTABLE or ACCUMULATOR quantity type
         if (changeQuantity > 0) inValue.setQuantity(changeQuantity).some else None
