+++ brambl-sdk/src/main/scala/co/topl/brambl/builders/UserInputValidations.scala
@@ -2,7 +2,7 @@ package co.topl.brambl.builders
 
 import cats.data.{Chain, NonEmptyChain, Validated, ValidatedNec}
 import cats.implicits.{catsSyntaxEitherId, catsSyntaxOptionId, catsSyntaxValidatedIdBinCompat0, toFoldableOps}
-import co.topl.brambl.models.box.{AssetMintingStatement, FungibilityType}
+import co.topl.brambl.models.box.{AssetMintingStatement, FungibilityType, QuantityDescriptorType}
 import co.topl.brambl.models.{LockAddress, SeriesId, TransactionOutputAddress}
 import co.topl.brambl.models.box.Value._
 import quivr.models.Int128
@@ -125,47 +125,24 @@ object UserInputValidations {
       UserInputError(s"Not enough LVLs in input to satisfy fee")
     )
 
-  // TODO: Finish implementing
-  // if agg type is immutable: then
-  // hmm undefined case: If 2 inputs with same type identity, immtable q=1, immutable q=2, we want to transfer only 1 of them
-  def validTransferQuantityDescriptors(
-    aggregateType: AggregateIdentifier,
-    testValues:    Seq[Value]
-  ): ValidatedNec[UserInputError, Unit] = ??? // {
-//    val groupedValues = testValues.groupBy(_.typeIdentifier.aggregateIdentifier)
-//    val toTransfer = groupedValues.get(aggregateType).map(_.map(_.quantity).sum).getOrElse(0)
-//    Validated.condNec(
-//      testValues.filter(_.isLvl).map(_.quantity: BigInt).sum >= fee + transferRequirements,
-//      (),
-//      UserInputError(s"Not enough LVLs in input to satisfy fee")
-//    )
-//  }
-
-  // TODO: Remove these
-  // The following two validations are temporary until we support all fungibility types
-  def fungibilityType(testType: Option[FungibilityType]): ValidatedNec[UserInputError, Unit] =
+  // TODO: The following validations are temporary until we support all quanitity descriptor types
+  // We currently only support LIQUID
+  def quantityDescriptorType(testType: Option[QuantityDescriptorType]): ValidatedNec[UserInputError, Unit] =
     Validated.condNec(
-      testType.isEmpty || testType.get == FungibilityType.GROUP_AND_SERIES,
+      testType.isEmpty || testType.get == QuantityDescriptorType.LIQUID,
       (),
-      UserInputError(s"Unsupported fungibility type. We currently only support GROUP_AND_SERIES")
+      UserInputError(s"Unsupported quantity descriptor type. We currently only support LIQUID")
     )
 
-  def identifierFungibilityType(testType: ValueTypeIdentifier): ValidatedNec[UserInputError, Unit] =
-    Validated.condNec(
-      testType match {
-        case GroupFungible(_, _, _) | SeriesFungible(_, _, _) => false
-        case _                                                => true
-      },
-      (),
-      UserInputError(s"Unsupported fungibility type. We currently only support GROUP_AND_SERIES")
-    )
+  def identifierQuantityDescriptorType(testType: ValueTypeIdentifier): ValidatedNec[UserInputError, Unit] =
+    quantityDescriptorType(testType.aggregateIdentifier.aggregateType.map(_.qdType))
 
-  def allValidFungibilityTypes(testValues: Seq[Value]): ValidatedNec[UserInputError, Unit] =
+  def allValidQuantityDescriptorTypes(testValues: Seq[Value]): ValidatedNec[UserInputError, Unit] =
     Validated.condNec(
-      // Any asset tokens must have valid fungibility
-      testValues.forall(v => !v.isAsset || v.asset.get.fungibility == FungibilityType.GROUP_AND_SERIES),
+      // Any asset tokens must have valid QuantityDescriptorType
+      testValues.map(_.typeIdentifier).map(identifierQuantityDescriptorType).forall(_.isValid),
       (),
-      UserInputError(s"All asset tokens must have valid fungibility type. We currently only support GROUP_AND_SERIES")
+      UserInputError(s"All asset tokens must have valid QuantityDescriptorType. We currently only support Liquid")
     )
 
   object TransactionBuilder {
@@ -183,10 +160,9 @@ object UserInputValidations {
         positiveQuantity(Some(amount), "quantity to transfer"),
         allInputLocksMatch(txos.map(_.transactionOutput.address), fromLockAddr, "fromLockAddr"),
         validTransferSupply(amount, transferValues),
-        identifierFungibilityType(transferIdentifier),
-        allValidFungibilityTypes(allValues),
+        identifierQuantityDescriptorType(transferIdentifier),
+        allValidQuantityDescriptorTypes(allValues),
         validFee(fee, allValues, if (transferIdentifier == LvlType) amount else 0)
-//        validTransferQuantityDescriptors(transferIdentifier.aggregateIdentifier, transferValues)
       ).fold.toEither
     } match {
       case Success(value) => value
@@ -257,7 +233,7 @@ object UserInputValidations {
           groupTxo.transactionOutput.value.value.group.flatMap(_.fixedSeries),
           seriesTxo.transactionOutput.value.value.series.map(_.seriesId)
         ),
-        fungibilityType(seriesTxo.transactionOutput.value.value.series.map(_.fungibility))
+        quantityDescriptorType(seriesTxo.transactionOutput.value.value.series.map(_.quantityDescriptor))
       ).fold.toEither
 
   }
